almost there

+++ CPP-Module02/ex03/bsp.cpp
@@ -1,7 +1,7 @@
 #include "Point.hpp"
 
 float area(Point const p1, Point const p2, Point const p3) {
-    return fabs((p1.getX() * (p2.getY() - p3.getY()) + p2.getX() * (p3.getY() - p1.getY()) + p3.getX() *(p1.getY() - p2.getY())) / 2.0);
+    return fabs((p1.getX().toFloat() * (p2.getY().toFloat() - p3.getY().toFloat()) + p2.getX().toFloat() * (p3.getY().toFloat() - p1.getY().toFloat()) + p3.getX().toFloat() *(p1.getY().toFloat() - p2.getY().toFloat())) / 2.0f);
 }
 
 bool bsp(Point const a, Point const b, Point const c, Point const point) {
